Added tester.py - a simple preset tester
Created a simple program that tests presets generated by the Scrya program in order to bugfix and find potential issues with the code. You simply enter your WLED ip and paste a json command into the command line as a single line command and it should work seamlessly. The alternative is to create a preset in WLED every time you want to test something, this is much faster.