fixed offset bug for rgb page, improved visibility
The old screen was split into 2 separate sections, because in the past grouping was not automatically detected and had to be manually inserted by the end user. Since we removed that functionality, I went ahead and removed the split screen and centered the IP address input box in the center of the screen with a small decoration to match.